requirements.txt
pip>=25.2
python-telegram-bot==22.3
aiohttp==3.9.5
requests==2.32.3
python-dotenv==1.0.1
yt-dlp==2025.07.21
imageio[ffmpeg]==2.36.1
pyrogram==2.0.106
tgcrypto==1.2.5
beautifulsoup4==4.12.3
# ffmpeg must be installed on the host machine (system package).
